* restructure TF operator library for building wheels
