+++ tensorflow_recordio/python/test_data/test_data_with_snappy_compression
